Create Task 1 (Number Game)
This is a Number Game which is my 1st Task of Codsoft. In this we can take any number between 1 to 100 and the system has to guess whether the number is correct or not. The guessed number tells whether its too low or too high.In this i have given 5 chances to guess the number. If you will be unable to guess then the number will be highlighted at the last.